[Python] Bing background picture

+++ Python-Practice/Wallpaper/main.py
@@ -10,14 +10,16 @@ for line in webpage:
 webpage_json = json.loads(webpage_str)#decode rss string
 date = webpage_json['images'][0]['startdate']#get date
 dest = "/home/penghou/Pictures/BingWallpaper/"+date+".jpg"#construct destination string to save the file
-url = "http://www.bing.com"
-url += webpage_json['images'][0]['url']#construct image url
-url = url.replace("1366","1920")#change url to fetch high resolution image
-url = url.replace("768","1080")#change url to fetch high resolution image
-urllib.urlretrieve(url,dest)
-command = "gsettings set org.gnome.desktop.background picture-uri 'file://" + dest +"'"#construct command of changing linux background
-start = time.time()
-while not os.path.isfile(dest):
-	if (time.time() - start) > 10:
-		break
-os.system(command)#execute command to change the desktop
+
+if not os.path.isfile(dest):#if the newest file is not exist, download and set wallpaper
+	url = "http://www.bing.com"
+	url += webpage_json['images'][0]['url']#construct image url
+	url = url.replace("1366","1920")#change url to fetch high resolution image
+	url = url.replace("768","1080")#change url to fetch high resolution image
+	urllib.urlretrieve(url,dest)
+	command = "gsettings set org.gnome.desktop.background picture-uri 'file://" + dest +"'"#construct command of changing linux background
+	start = time.time()
+	while not os.path.isfile(dest):
+		if (time.time() - start) > 10:
+			break
+	os.system(command)#execute command to change the desktop
